+++ ec/__main__.py
@@ -25,42 +25,21 @@ def main():
     
     target_path = argv.pop(0)
     
+    sys.argv = sys.argv[:1] + argv # alter sys.argv so that the modules could process them
+    
     if path.isfile(target_path):
       core.start(load_module(target_path), argv)
       
     elif path.isdir(target_path): # launch the dir with its children as groups
-      import imp
       from glob import glob
-      from collections import OrderedDict
       
-      from modules.classes import Group
-      
-      Container = imp.new_module(target_path)
-      Members = OrderedDict()
+      from ec import member # ec has to be imported in order to make the import hook work
       
       for Module in [load_module(module_path) for module_path in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(module_path)]:
         __ec_member__ = getattr(Module, '__ec_member__', None)
+        
         if __ec_member__: # register only modules that are designed for ec
-          __ec_member__.__load_members__()
-          name = Module.__name__
-          print Module, name, '--------'
-          setattr(Container, name, Module)
-          Members[name] = __ec_member__
-      
-      BaseGroup = Group(Container, {'name': target_path}) # brand the Container
-      BaseGroup.Config['Members'] = Members
-      
-      '''
-      class Container: # a class to emulate a module
-        def __init__(self, Modules):
-          for Module in Modules:
-            if hasattr(Module, '__ec_member__'): # register only modules that are designed for ec
-              setattr(self, Module.__name__, Module)
-      Container = Container()
-      '''
-      
-      BaseGroup.__load_members__()
-      core.start(Container, argv)
+          member(Module)
       
     else:
       show_usage()  
@@ -70,3 +49,4 @@ def show_usage():
 
 if __name__ == '__main__':
   main()
+