+++ tests/support/nester.py
@@ -4,9 +4,13 @@ nester
 
 Used to test nested modules.
 """
-from ec.ec import group
+from ec.ec import member, task
 
 import target_script
 
-group(target_script)
+member(target_script) # add the imported member to the script
 
+@task
+def task1():
+  print 'task1'
+