+++ tests/test_nested.py
@@ -12,8 +12,24 @@ from support.helpers import shell_exec
 # Modifications to test_dispatch
 import test_dispatch
 
-test_dispatch.script_name = 'support/nester.py'
-test_dispatch.command_prefix = 'target_script/'
+def dispatch(argStr='', input='', flag=''):
+  """Dispatches command to the target script.
+  """
+  if flag == '-h':
+    rest = flag
+    
+  elif flag == '-p':
+    
+    rest = '%s target_script/%s' % (flag, argStr)
+    
+  else:
+    
+    rest = 'target_script/%s' % argStr
+    
+  return shell_exec('python tests/support/nester.py %s' % rest, input=input)
+    
+
+test_dispatch.dispatch = dispatch
 
 TestNested = test_dispatch.TestDispatch